Improve loading UserInfo and Groups
The extension only worked either with groups or with userinfo.
With this patch, in the first step the group is taken from $wgSimpleSAMLphp_GroupMap. If this is found, the check in next groups will not overwrite the already found groups.

This change was reported by a third party, internal reference id: ERM14518

Change-Id: Idfb63a51ff8e22f5755d860b15b6fe8582ee4a6c